Added basic monsters that notice one turn every turn. Also added enums for game states

+++ src/map.rs
@@ -148,7 +148,7 @@ impl Map {
 
 impl BaseMap for Map {
     fn is_opaque(&self, idx: usize) -> bool {
-        self.tiles[idx as usize] == TileType::Wall
+        self.tiles[idx] == TileType::Wall
     }
 }
 impl Algorithm2D for Map {
